requirements.txt
langchain-community
langchain_groq
langchain_huggingface
sentence_transformers
streamlit
beautifulsoup4
ddgs
numexpr